Rename generic Edit and Preview methods and file names to github.
