+++ frontend/src/Components/GithubEditConditions.js
@@ -1,10 +1,77 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+import axios from '../HelperFunctions/customAxios';
+import React, { useEffect, useState } from 'react';
+import PropTypes from 'prop-types'
+import { toast } from 'react-toastify';
 
-export default function PreviewConditions({ modalData, showPreview, setShowPreview }) {
+export default function GithubEditConditions({roomId, modalData, setModalData, roomConditions, setRoomConditions, showGithubEditable, setShowGithubEditable}) {
+    const [currentData, setCurrentData] = useState({});
+
+    function handleCheckboxChange(e){
+        const key = e.target.name;
+        setCurrentData((previousData) => {
+            return {...previousData, [key]: !previousData[key]};
+        });
+    }
+
+    function handleClose(){
+        setCurrentData(modalData.data ? {...modalData.data} : {});
+        setShowGithubEditable(false);
+    }
+
+    function handleSave(){
+
+        async function saveData(roomId, modalData, currentData){
+            const thirdPartyAccount = modalData.thirdPartyAccount.toLowerCase();
+            const conditionType = modalData.conditionType.toLowerCase();
+            const ownerType = modalData.type;
+
+            let url;
+            if(thirdPartyAccount === 'github') {
+                url = `/api/rooms/${roomId}/github/${ownerType}/${conditionType}`;
+            }
+
+            // Excludes the 'key' and updates the 'data' property.
+            let dataToBeSent = {
+                "type": ownerType,
+                "third_party_account": thirdPartyAccount,
+                "owner": modalData.owner,
+                "condition_type": conditionType,
+                "data": currentData
+            }
+
+            return axios.put(url, dataToBeSent);
+        }
+
+        const resp = saveData(roomId, modalData, currentData);
+
+        resp.then( (resp) => {
+            setModalData(previousData => ({...previousData, data: currentData}));
+
+            setRoomConditions(previousData => {
+                return previousData.map(item => {
+                    if (item.key === modalData.key) {
+                        return {...modalData, data: currentData};
+                    }
+                    else {
+                        return item;
+                    }
+                });
+            });
+            toast.success("Succesfully updated!");
+        })
+        .catch( () => {
+            toast.error("Failed to save data");
+        })
+        handleClose();
+    }
+
+    useEffect( () => {
+        setCurrentData(modalData.data ? {...modalData.data} : {});
+    }, [modalData]);
+
     return (
-        <>
-        {showPreview ? (
+    <>
+        {showGithubEditable ? (
             <>
             <div
                 className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none"
@@ -27,7 +94,7 @@ export default function PreviewConditions({ modalData, showPreview, setShowPrevi
 
                         <button
                             className="p-1 ml-auto"
-                            onClick={() => setShowPreview(false)}
+                            onClick={handleClose}
                         >
                             <svg
                             className="w-6 h-6"
@@ -69,7 +136,7 @@ export default function PreviewConditions({ modalData, showPreview, setShowPrevi
 
                         <div className='p-2 w-full'>
                             { modalData.data
-                                ? Object.entries(modalData.data).map( ([key, value]) => {
+                                ? Object.entries(currentData).map( ([key, value]) => {
                                     return (
                                         <div key={key}>
                                             <input
@@ -77,15 +144,32 @@ export default function PreviewConditions({ modalData, showPreview, setShowPrevi
                                             type="checkbox"
                                             id={`checkbox-${key}`}
                                             name={key}
-                                            disabled={true}
                                             checked={value}
+                                            onChange={(e) => handleCheckboxChange(e)}
                                             />
-                                            <label htmlFor={`checkbox-${key}`}>{key}</label>
+                                            <label className='w-full' htmlFor={`checkbox-${key}`}>{key}</label>
                                         </div>
                                     )
                                 })
                                 : <></>
                             }
+                        </div>
+
+
+                        <div className='flex justify-end items-center mt-3'>
+                            <button
+                            className="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none font-medium rounded-lg text-sm px-4 py-2 text-center mr-2"
+                            onClick={handleSave}
+                            >
+                                Save
+                            </button>
+
+                            <button
+                            className="text-white bg-red-700 hover:bg-red-800 focus:outline-none font-medium rounded-lg text-sm px-4 py-2 text-center "
+                            onClick={handleClose}
+                            >
+                                Cancel
+                            </button>
                         </div>
 
                     </div>
@@ -95,12 +179,16 @@ export default function PreviewConditions({ modalData, showPreview, setShowPrevi
             <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
             </>
         ) : null}
-        </>
-    );
+    </>
+  )
 }
 
-PreviewConditions.propTypes = {
+GithubEditConditions.propTypes = {
+    roomId: PropTypes.string,
     modalData: PropTypes.object,
-    showPreview: PropTypes.bool,
-    setShowPreview: PropTypes.func
+    setModalData: PropTypes.func,
+    roomConditions: PropTypes.array,
+    setRoomConditions: PropTypes.func,
+    showGithubEditable: PropTypes.bool,
+    setShowGithubEditable: PropTypes.func
 }
