+++ frontend/src/Components/GithubPreviewConditions.js
@@ -1,77 +1,10 @@
-import axios from '../HelperFunctions/customAxios';
-import React, { useEffect, useState } from 'react';
-import PropTypes from 'prop-types'
-import { toast } from 'react-toastify';
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 
-export default function EditConditions({roomId, modalData, setModalData, roomConditions, setRoomConditions, showEditable, setShowEditable}) {
-    const [currentData, setCurrentData] = useState({});
-
-    function handleCheckboxChange(e){
-        const key = e.target.name;
-        setCurrentData((previousData) => {
-            return {...previousData, [key]: !previousData[key]};
-        });
-    }
-
-    function handleClose(){
-        setCurrentData(modalData.data ? {...modalData.data} : {});
-        setShowEditable(false);
-    }
-
-    function handleSave(){
-
-        async function saveData(roomId, modalData, currentData){
-            const thirdPartyAccount = modalData.thirdPartyAccount.toLowerCase();
-            const conditionType = modalData.conditionType.toLowerCase();
-            const ownerType = modalData.type;
-
-            let url;
-            if(thirdPartyAccount === 'github') {
-                url = `/api/rooms/${roomId}/github/${ownerType}/${conditionType}`;
-            }
-
-            // Excludes the 'key' and updates the 'data' property.
-            let dataToBeSent = {
-                "type": ownerType,
-                "third_party_account": thirdPartyAccount,
-                "owner": modalData.owner,
-                "condition_type": conditionType,
-                "data": currentData
-            }
-
-            return axios.put(url, dataToBeSent);
-        }
-
-        const resp = saveData(roomId, modalData, currentData);
-
-        resp.then( (resp) => {
-            setModalData(previousData => ({...previousData, data: currentData}));
-
-            setRoomConditions(previousData => {
-                return previousData.map(item => {
-                    if (item.key === modalData.key) {
-                        return {...modalData, data: currentData};
-                    }
-                    else {
-                        return item;
-                    }
-                });
-            });
-            toast.success("Succesfully updated!");
-        })
-        .catch( () => {
-            toast.error("Failed to save data");
-        })
-        handleClose();
-    }
-
-    useEffect( () => {
-        setCurrentData(modalData.data ? {...modalData.data} : {});
-    }, [modalData]);
-
+export default function GithubPreviewConditions({ modalData, showGithubPreview, setShowGithubPreview }) {
     return (
-    <>
-        {showEditable ? (
+        <>
+        {showGithubPreview ? (
             <>
             <div
                 className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none"
@@ -94,7 +27,7 @@ export default function EditConditions({roomId, modalData, setModalData, roomCon
 
                         <button
                             className="p-1 ml-auto"
-                            onClick={handleClose}
+                            onClick={() => setShowGithubPreview(false)}
                         >
                             <svg
                             className="w-6 h-6"
@@ -136,7 +69,7 @@ export default function EditConditions({roomId, modalData, setModalData, roomCon
 
                         <div className='p-2 w-full'>
                             { modalData.data
-                                ? Object.entries(currentData).map( ([key, value]) => {
+                                ? Object.entries(modalData.data).map( ([key, value]) => {
                                     return (
                                         <div key={key}>
                                             <input
@@ -144,32 +77,15 @@ export default function EditConditions({roomId, modalData, setModalData, roomCon
                                             type="checkbox"
                                             id={`checkbox-${key}`}
                                             name={key}
+                                            disabled={true}
                                             checked={value}
-                                            onChange={(e) => handleCheckboxChange(e)}
                                             />
-                                            <label className='w-full' htmlFor={`checkbox-${key}`}>{key}</label>
+                                            <label htmlFor={`checkbox-${key}`}>{key}</label>
                                         </div>
                                     )
                                 })
                                 : <></>
                             }
-                        </div>
-
-
-                        <div className='flex justify-end items-center mt-3'>
-                            <button
-                            className="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none font-medium rounded-lg text-sm px-4 py-2 text-center mr-2"
-                            onClick={handleSave}
-                            >
-                                Save
-                            </button>
-
-                            <button
-                            className="text-white bg-red-700 hover:bg-red-800 focus:outline-none font-medium rounded-lg text-sm px-4 py-2 text-center "
-                            onClick={handleClose}
-                            >
-                                Cancel
-                            </button>
                         </div>
 
                     </div>
@@ -179,16 +95,12 @@ export default function EditConditions({roomId, modalData, setModalData, roomCon
             <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
             </>
         ) : null}
-    </>
-  )
+        </>
+    );
 }
 
-EditConditions.propTypes = {
-    roomId: PropTypes.string,
+GithubPreviewConditions.propTypes = {
     modalData: PropTypes.object,
-    setModalData: PropTypes.func,
-    roomConditions: PropTypes.array,
-    setRoomConditions: PropTypes.func,
-    showEditable: PropTypes.bool,
-    setShowEditable: PropTypes.func
+    showGithubPreview: PropTypes.bool,
+    setShowGithubPreview: PropTypes.func
 }
